Fix of __repr__ and add_rotation function
Rotation must include a center. Verified that translations work correct now.
Note that elliptical arcs are applied correctly! So comment "# Todo transformations aren't applied correctly to elliptical arcs" (_path.py) is incorrect.
A note on comment
    # ToDo apply transformation beforehand (in Path) for consistency with other geometric objects. If you (the reader)
    #  know how to easily apply an affine transformation to an ellipse feel free to make a pull request.
(_elliptical_arc.py): it seems to me that the 'rotation' parameter of the ellipticalarc interferes with a transformation that results in rotation. To get it right the 'rotation' part of the translation must be extracted.